addTableのフック化
