+++ booking_web/src/components/atoms/button/AddTableRelationButton.jsx
@@ -1,22 +1,13 @@
 import React, { useState } from "react";
-import { useSetRecoilState } from "recoil";
 import { Box, Button, Select } from "@chakra-ui/react";
 
-import { addTableRelation, getAllBooking } from "../../../utils/api";
-import { today } from "../../../utils/date";
-import {
-  allBookingState,
-  todayBookingState,
-  noAssignedBookingState,
-} from "../../../atom/state";
-import { useMessage } from "../../../hooks/useMessage";
+
+
+import { useAddTableRelation } from "../../../hooks/useAddTabelRelation";
 
 export const AddTableRelationButton = (props) => {
   const { bookingId } = props;
-  const setAllBooking = useSetRecoilState(allBookingState);
-  const setTodayBooking = useSetRecoilState(todayBookingState);
-  const setNoAssigendBooking = useSetRecoilState(noAssignedBookingState);
-  const { showMessage } = useMessage();
+  const { addTable } = useAddTableRelation();
 
 
   const tableNum = [
@@ -33,34 +24,15 @@ export const AddTableRelationButton = (props) => {
     { id: 21, name: "21番" },
   ];
 
-  const [table, setTable] = useState(1);
+  const [tableId, setTableId] = useState(1);
 
-  const onClickAddTableRelation = async () => {
-    const data = {
-      booking: {
-        id: bookingId,
-      },
-      table: {
-        id: table,
-      },
-    };
-    await addTableRelation(data);
-    const allBooking = await getAllBooking();
-    setAllBooking(allBooking);
-    const todayBooking = allBooking.filter(
-      (booking) => booking.date === today()
-    );
-    setTodayBooking(todayBooking);
-    const noAssigendBooking = todayBooking.filter((booking) => {
-      return booking.tables.length === 0;
-    });
-    setNoAssigendBooking(noAssigendBooking);
-    showMessage({ title: "席を割り振りました", status: "success" });
-  };
+  const onClickAddTableRelation = () => {
+    addTable(bookingId, tableId);
+  }
 
   return (
     <Box display={"flex"}>
-      <Select bg={"white"} w={"100px"} value={table} onChange={(e) => setTable(e.target.value)}>
+      <Select bg={"white"} w={"100px"} value={tableId} onChange={(e) => setTableId(e.target.value)}>
         {tableNum.map((table) => (
           <option key={table.id} value={table.id}>
             {table.name}
